Move Fake path based providers to FakeV@Provider
